+++ src/lib/services/tx/rest.ts
@@ -7,51 +7,63 @@ import {
   zTxsByAddressResponseRest,
   zTxsByHashResponseRest,
 } from "../types";
+import { queryWithArchivalFallback } from "../utils";
 
-export const getTxDataRest = async (endpoint: string, txHash: string) =>
-  axios
-    .get(`${endpoint}/cosmos/tx/v1beta1/txs/${encodeURI(txHash)}`)
-    .then(({ data }) => parseWithError(zTxByHashResponseRest, data));
+export const getTxDataRest = (endpoint: string, txHash: string) => {
+  const fetch = (endpoint: string) =>
+    axios
+      .get(`${endpoint}/cosmos/tx/v1beta1/txs/${encodeURI(txHash)}`)
+      .then(({ data }) => parseWithError(zTxByHashResponseRest, data));
 
-export const getTxsByHashRest = async (endpoint: string, txHash: string) =>
+  return queryWithArchivalFallback(endpoint, fetch);
+};
+
+export const getTxsByHashRest = (endpoint: string, txHash: string) =>
   axios
     .get(`${endpoint}/cosmos/tx/v1beta1/txs/${encodeURI(txHash)}`)
     .then(({ data }) => parseWithError(zTxsByHashResponseRest, data));
 
-export const getTxsByContractAddressRest = async (
+export const getTxsByContractAddressRest = (
   endpoint: string,
   contractAddress: BechAddr32,
   limit: number,
   offset: number
-) =>
-  Promise.allSettled([
-    axios.get(`${endpoint}/cosmos/tx/v1beta1/txs`, {
-      params: {
-        order_by: 2,
-        limit,
-        page: offset / limit + 1,
-        query: `wasm._contract_address='${encodeURI(contractAddress)}'`,
-      },
-    }),
-    axios.get(`${endpoint}/cosmos/tx/v1beta1/txs`, {
-      params: {
-        order_by: 2,
-        limit,
-        page: offset / limit + 1,
-        events: `wasm._contract_address='${encodeURI(contractAddress)}'`,
-      },
-    }),
-  ]).then(([queryParam, eventsParam]) => {
-    if (queryParam.status === "fulfilled")
-      return parseWithError(zTxsByAddressResponseRest, queryParam.value.data);
+) => {
+  const fetch = (endpoint: string) =>
+    Promise.allSettled([
+      axios.get(`${endpoint}/cosmos/tx/v1beta1/txs`, {
+        params: {
+          order_by: 2,
+          limit,
+          page: offset / limit + 1,
+          query: `wasm._contract_address='${encodeURI(contractAddress)}'`,
+        },
+      }),
+      axios.get(`${endpoint}/cosmos/tx/v1beta1/txs`, {
+        params: {
+          order_by: 2,
+          limit,
+          page: offset / limit + 1,
+          events: `wasm._contract_address='${encodeURI(contractAddress)}'`,
+        },
+      }),
+    ]).then(([queryParam, eventsParam]) => {
+      if (queryParam.status === "fulfilled")
+        return parseWithError(zTxsByAddressResponseRest, queryParam.value.data);
 
-    if (eventsParam.status === "fulfilled")
-      return parseWithError(zTxsByAddressResponseRest, eventsParam.value.data);
+      if (eventsParam.status === "fulfilled")
+        return parseWithError(
+          zTxsByAddressResponseRest,
+          eventsParam.value.data
+        );
 
-    throw new Error("No data found (getTxsByContractAddressRest)");
-  });
+      throw new Error("No data found (getTxsByContractAddressRest)");
+    });
 
-export const getTxsByAccountAddressRest = async (
+  return queryWithArchivalFallback(endpoint, fetch);
+};
+
+export const getTxsByAccountAddressRest = (
   endpoint: string,
   address: BechAddr20,
   limit: number,