+++ CHANGELOG.md
@@ -39,6 +39,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Features
 
+- [#1280](https://github.com/alleslabs/celatone-frontend/pull/1280) Support archival fallback for txs and blocks
+
 ### Improvements
 
 - [#1281](https://github.com/alleslabs/celatone-frontend/pull/1281) Add advanced options helper text, limit fetching verification infos to hex address only, add more module information